Refactor minigame game implementations to package
