+++ minigames/games/__init__.py
@@ -1,0 +1,12 @@
+from django.conf import settings
+from django.utils.module_loading import import_string
+
+from minigames.games.base import BaseGame, GameScore
+from minigames.games.first_to_n import FirstToN
+from minigames.games.lockout_bingo import LockoutBingo
+
+game_registry: dict[str, BaseGame] = {
+    name: import_string(path)() for name, path in settings.MINIGAMES.items()
+}
+
+__all__ = ["BaseGame", "GameScore", "FirstToN", "LockoutBingo", "game_registry"]
