+++ minigames/games/base.py
@@ -1,0 +1,57 @@
+from abc import ABC, abstractmethod
+from datetime import datetime
+from typing import NamedTuple
+
+
+class GameScore(NamedTuple):
+    id: int
+    player_id: int
+    team_id: int
+    score_id: int
+    points: float
+
+    score_count_300: int
+    score_count_100: int
+    score_count_50: int
+    score_count_miss: int
+    score_best_combo: int
+    score_perfect: bool
+    score_mods_json: dict
+    score_accuracy: float
+    score_rank: str
+    score_date: datetime
+
+    beatmap_id: int
+    beatmap_creator_name: str
+    beatmap_status: int
+
+    score_performance_total: float | None = None
+    score_difficulty_total: float | None = None
+
+
+class BaseGame(ABC):
+    @property
+    @abstractmethod
+    def game_type(self) -> str: ...
+
+    @property
+    @abstractmethod
+    def display_name(self) -> str: ...
+
+    def get_settings(self, data: dict) -> dict:
+        game_length = int(data.get("game_length", 60 * 60))
+        return {"game_length": min(game_length, 60 * 60 * 2)}
+
+    def get_initial_state(self, config: dict) -> dict:
+        return {}
+
+    def process_scores(
+        self, scores: list[GameScore], config: dict, initial_state: dict
+    ) -> dict:
+        return {
+            "state": {},
+            "win_condition_reached": False,
+            "players": {},
+            "teams": {},
+            "scores": {},
+        }
